+++ accounting/src/apideck_accounting_unify/billpayments.py
@@ -120,7 +120,7 @@ class BillPayments(BaseSDK):
         )
 
         def next_func() -> Optional[models.AccountingBillPaymentsAllResponse]:
-            body = utils.unmarshal_json(http_res.text, Dict[Any, Any])
+            body = utils.unmarshal_json(http_res.text, Union[Dict[Any, Any], List[Any]])
             next_cursor = JSONPath("$.meta.cursors.next").parse(body)
 
             if len(next_cursor) == 0:
@@ -315,7 +315,7 @@ class BillPayments(BaseSDK):
         )
 
         def next_func() -> Optional[models.AccountingBillPaymentsAllResponse]:
-            body = utils.unmarshal_json(http_res.text, Dict[Any, Any])
+            body = utils.unmarshal_json(http_res.text, Union[Dict[Any, Any], List[Any]])
             next_cursor = JSONPath("$.meta.cursors.next").parse(body)
 
             if len(next_cursor) == 0: