+++ accounting/src/apideck_accounting_unify/locations.py
@@ -118,7 +118,7 @@ class Locations(BaseSDK):
         )
 
         def next_func() -> Optional[models.AccountingLocationsAllResponse]:
-            body = utils.unmarshal_json(http_res.text, Dict[Any, Any])
+            body = utils.unmarshal_json(http_res.text, Union[Dict[Any, Any], List[Any]])
             next_cursor = JSONPath("$.meta.cursors.next").parse(body)
 
             if len(next_cursor) == 0:
@@ -310,7 +310,7 @@ class Locations(BaseSDK):
         )
 
         def next_func() -> Optional[models.AccountingLocationsAllResponse]:
-            body = utils.unmarshal_json(http_res.text, Dict[Any, Any])
+            body = utils.unmarshal_json(http_res.text, Union[Dict[Any, Any], List[Any]])
             next_cursor = JSONPath("$.meta.cursors.next").parse(body)
 
             if len(next_cursor) == 0: