+++ accounting/src/apideck_accounting_unify/logs.py
@@ -6,7 +6,7 @@ from apideck_accounting_unify._hooks import HookContext
 from apideck_accounting_unify.types import OptionalNullable, UNSET
 from apideck_accounting_unify.utils import get_security_from_env
 from jsonpath import JSONPath
-from typing import Any, Dict, Mapping, Optional, Union
+from typing import Any, Dict, List, Mapping, Optional, Union
 
 
 class Logs(BaseSDK):
@@ -103,7 +103,7 @@ class Logs(BaseSDK):
         )
 
         def next_func() -> Optional[models.VaultLogsAllResponse]:
-            body = utils.unmarshal_json(http_res.text, Dict[Any, Any])
+            body = utils.unmarshal_json(http_res.text, Union[Dict[Any, Any], List[Any]])
             next_cursor = JSONPath("$.meta.cursors.next").parse(body)
 
             if len(next_cursor) == 0:
@@ -277,7 +277,7 @@ class Logs(BaseSDK):
         )
 
         def next_func() -> Optional[models.VaultLogsAllResponse]:
-            body = utils.unmarshal_json(http_res.text, Dict[Any, Any])
+            body = utils.unmarshal_json(http_res.text, Union[Dict[Any, Any], List[Any]])
             next_cursor = JSONPath("$.meta.cursors.next").parse(body)
 
             if len(next_cursor) == 0: