+++ accounting/src/apideck_accounting_unify/models/accounting_suppliersupdateop.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 from .httpmetadata import HTTPMetadata, HTTPMetadataTypedDict
-from .supplier_input import SupplierInput, SupplierInputTypedDict
+from .supplier import SupplierInput, SupplierInputTypedDict
 from .unexpectederrorresponse import (
     UnexpectedErrorResponse,
     UnexpectedErrorResponseTypedDict,