+++ accounting/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [python v0.11.1] accounting
 ### Releases
 - [PyPI v0.11.1] https://pypi.org/project/apideck-accounting-unify/0.11.1 - accounting
+
+## 2025-04-22 00:04:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.537.0 (2.588.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.12.0] accounting
+### Releases
+- [PyPI v0.12.0] https://pypi.org/project/apideck-accounting-unify/0.12.0 - accounting